requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
python-multipart==0.0.27
networkx==3.6.1
boto3==1.42.79
graphifyy==0.5.0
pdfplumber==0.11.9
rdflib==7.6.0
pyyaml==6.0.3
requests==2.32.3
click==8.3.1
python-slugify==8.0.4
python-dotenv==1.1.0
